restructure the Model  into src/Model
